백엔드 Resolver명과 프론트 Component명 동일화 & searchItems Header에서 분리하여 Component화
