+++ src/App.js
@@ -9,18 +9,18 @@ import { ApolloProvider, useReactiveVar } from '@apollo/client';
 import { HelmetProvider } from 'react-helmet-async';
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom"
 import { ThemeProvider } from 'styled-components';
-import Login from './screens/auth/Login';
-import Home from './screens/main/Home';
+import Login from './screens/main/users/Login';
+import Home from './screens/main/items/Home';
 import { client, darkModeVar, isLoggedInVar } from './utils/apollo';
 import { darkTheme, GlobalStyles, lightTheme } from './utils/styles';
-import SignUp from './screens/auth/SignUp';
-import UserDetail from './screens/main/UserDetail';
+import SignUp from './screens/main/users/CreateUser';
+import UserDetail from './screens/main/users/SeeUser';
 import GetTokenFromUrl from './utils/GetTokenFromUrl';
-import UserEdit from './screens/main/UserEdit';
-import ItemEdit from './screens/main/ItemEdit';
-import ItemUpload from './screens/main/ItemUpload';
-import ItemDelete from './screens/main/ItemDelete';
-import ItemDetail from './screens/main/ItemDetail';
+import UserEdit from './screens/main/users/EditUser';
+import ItemEdit from './screens/main/items/EditItem';
+import ItemUpload from './screens/main/items/UploadItem';
+import ItemDelete from './screens/main/items/DeleteItem';
+import ItemDetail from './screens/main/items/SeeItem';
 import NotFound from './components/shared/utils/NotFound';
 
 function App() {
