+++ src/components/main/messages/CreateRoomBtn.js
@@ -27,7 +27,7 @@ const CREATE_ROOM = gql`
     }
     ${MESSAGE_DEFAULT_FRAGMENT}
 `
-export default function ChatBtn({ text, userId }) {
+export default function CreateRoomBtn({ text, userId }) {
 
     // MessageRoom을 on하기 위한 함수
     const enterRoom = (userId) => chatUserIdVar(userId)
@@ -83,7 +83,7 @@ export default function ChatBtn({ text, userId }) {
     )
 }
 
-ChatBtn.propTypes = {
+CreateRoomBtn.propTypes = {
     text: propTypes.string,
     userId: propTypes.number
 }
