+++ src/components/main/users/Header.js
@@ -1,7 +1,7 @@
 /* 
 작성자 : SJ
 작성일 : 2022.01.07
-수정일 : 2022.01.14
+수정일 : 2022.01.16
 */
 
 /*
@@ -20,13 +20,11 @@ import useLoggedInUser from '../../../hooks/useLoggedInUser'
 import { colors } from '../../../utils/styles'
 import FontAwesomeBtn from '../../shared/buttons/FontAwesomeBtn'
 import UserAvatar from './UserAvatar'
-import { gql, useLazyQuery, useReactiveVar } from "@apollo/client"
-import { darkModeVar, searchDataVar } from "../../../utils/apollo"
-import { useEffect, useState } from 'react'
-import Input from '../../shared/form/Input'
-import { useForm } from 'react-hook-form'
-import { ITEM_DISPLAY_FRAGMENT } from '../../shared/utils/fragments'
-import InputWithFontAwesome from '../../shared/form/InputWithFontAwesome'
+import { useReactiveVar } from "@apollo/client"
+import { darkModeVar, searchDataVar, searchModeVar } from "../../../utils/apollo"
+
+import SearchItems from '../items/SearchItems'
+import { useState } from 'react'
 
 const Container = styled.header`
     padding:20px 70px;
@@ -60,50 +58,16 @@ const Tabs = styled.div`
 const Tab = styled.button`
 `
 
-const SEARCH_ITEMS = gql`
-    query searchItems($keyword:String!){
-        searchItems(keyword:$keyword){
-            ...ItemDisplayFragment
-        }
-    }
-    ${ITEM_DISPLAY_FRAGMENT}
-`
 
 export default function Header() {
-    const [searchMode, setSearchMode] = useState(false)
+    const searchMode = useReactiveVar(searchModeVar)
     const searchData = useReactiveVar(searchDataVar)
     const darkMode = useReactiveVar(darkModeVar)
     const { loggedInUser } = useLoggedInUser()
     const history = useHistory()
 
-    // state를 이용하여 searchMode라면 Input창을, 아니라면 search 버튼을 띄워줄 것이다.
-    const getSearchMode = (bool) => setSearchMode(bool)
-
-
-    // searchItems 구현부
-    // searchItems로 받은 data를 Home으로 보내주어 사용할 것이다.
-    const { register, handleSubmit, formState, setValue } = useForm({
-        mode: "onChange"
-    })
-
-    // searchItems 성공 시 searchDataVar를 Reactive Variables로 data를 받아놓는다.
-    const searchCompleted = ({ searchItems }) => {
-        searchDataVar(searchItems)
-        history.push("/")
-    }
-    const [searchItems, { loading }] = useLazyQuery(SEARCH_ITEMS, {
-        onCompleted: searchCompleted
-    })
-    const onValid = ({ keyword }) => {
-        if (loading) return;
-
-        searchItems({
-            variables: {
-                keyword
-            }
-        })
-        setValue("keyword", "")
-    }
+    // Reactive Variables를 이용하여 searchMode라면 Input창을, 아니라면 search 버튼을 띄워줄 것이다.
+    const getSearchMode = (bool) => searchModeVar(bool)
 
     // path를 인자로 받아 원하는 path로 보내준다.
     const sendWhere = (path) => history.push(path)
@@ -140,25 +104,7 @@ export default function Header() {
                         />
                     ) : (
                         searchMode ? (
-                            <InputWithFontAwesome
-                                onClick={formState.isValid && searchMode ? handleSubmit(onValid) : () => getSearchMode(false)}
-                                icon={faSearch}
-                                color={darkMode ? colors.white : colors.black}
-                                size={"lg"}
-                                disabled={!formState.isValid || loading}
-                            >
-                                <Input
-                                    placeholder='동네 등 키워드'
-                                    {...register("keyword", {
-                                        required: true,
-                                        minLength: {
-                                            value: 2,
-                                            message: "두 글자 이상을 입력하세요."
-                                        }
-                                    })}
-                                    required
-                                />
-                            </InputWithFontAwesome>
+                            <SearchItems />
                         ) : (
                             <FontAwesomeBtn
                                 icon={faSearch}
