+++ src/screens/main/users/EditUser.js
@@ -11,19 +11,19 @@
 
 import { useForm } from 'react-hook-form'
 import { useHistory, useParams, useLocation } from 'react-router-dom'
-import InputError from '../../components/shared/form/InputError'
-import Input from '../../components/shared/form/Input'
-import useLoggedInUser from '../../hooks/useLoggedInUser'
-import Button from "../../components/shared/buttons/Button"
-import { kakaoLocationApi } from '../../apis/locationApi'
+import InputError from '../../../components/shared/form/InputError'
+import Input from '../../../components/shared/form/Input'
+import useLoggedInUser from '../../../hooks/useLoggedInUser'
+import Button from "../../../components/shared/buttons/Button"
+import { kakaoLocationApi } from '../../../apis/locationApi'
 import { useEffect, useState } from 'react'
 import { gql, useMutation } from '@apollo/client'
 import styled from 'styled-components'
-import { colors } from '../../utils/styles'
-import NotAuthorized from '../../components/shared/utils/NotAuthorized'
-import FormLayout from '../../components/layouts/FormLayout'
-import ItemPhoto from '../../components/main/items/ItemPhoto';
-import MainLayout from '../../components/layouts/MainLayout';
+import { colors } from '../../../utils/styles'
+import NotAuthorized from '../../../components/shared/utils/NotAuthorized'
+import FormLayout from '../../../components/layouts/FormLayout'
+import ItemPhoto from '../../../components/main/items/ItemPhoto';
+import MainLayout from '../../../components/layouts/MainLayout';
 
 const Container = styled.div`
     max-width:600px;
@@ -64,7 +64,7 @@ const EDIT_USER = gql`
         }
     }
 `
-export default function UserEdit() {
+export default function EditUser() {
     const { id } = useParams()
     const history = useHistory()
     const { state } = useLocation()
