+++ src/screens/main/users/Login.js
@@ -10,16 +10,16 @@ import { gql, useMutation } from '@apollo/client';
 import { useForm } from 'react-hook-form';
 import { useLocation } from 'react-router-dom';
 import styled from 'styled-components';
-import FormLayout from '../../components/layouts/FormLayout';
-import Button from '../../components/shared/buttons/Button';
-import Divider from '../../components/shared/utils/Divider';
-import SendAnywhere from '../../components/shared/utils/SendAnywhere';
-import { logUserIn } from '../../utils/apollo';
-import NaverLogin from '../../utils/NaverLogin';
-import { colors } from '../../utils/styles';
-import Input from '../../components/shared/form/Input';
-import InputError from '../../components/shared/form/InputError';
-import GithubLogin from '../../utils/GithubLogin';
+import FormLayout from '../../../components/layouts/FormLayout';
+import Button from '../../../components/shared/buttons/Button';
+import Divider from '../../../components/shared/utils/Divider';
+import SendAnywhere from '../../../components/shared/utils/SendAnywhere';
+import { logUserIn } from '../../../utils/apollo';
+import NaverLogin from '../../../utils/NaverLogin';
+import { colors } from '../../../utils/styles';
+import Input from '../../../components/shared/form/Input';
+import InputError from '../../../components/shared/form/InputError';
+import GithubLogin from '../../../utils/GithubLogin';
 
 const SocialLogins = styled.div`
     margin-top:5px;
@@ -33,9 +33,6 @@ const SocialLogins = styled.div`
 const Greeting = styled.span`
     color:${colors.green};
     text-align:center;
-`
-
-const GithubLoginBox = styled.div`
 `
 
 const LOGIN_MUTATION = gql`
